POC: to enable rust logs on python.
